+++ src/nmod_mpoly/set_term_coeff_ui.c
@@ -15,7 +15,7 @@
 void nmod_mpoly_set_term_coeff_ui(nmod_mpoly_t A, slong i, ulong c,
                                                     const nmod_mpoly_ctx_t ctx)
 {
-    if (i >= A->length)
+    if (i >= (ulong) A->length)
     {
         flint_throw(FLINT_ERROR, "nmod_mpoly_set_term_coeff_ui: index out of range.");
     }
